Added endpoints for registration

+++ participant-tracking/src/main/java/org/ikuven/bbut/tracking/repository/ParticipantRepository.java
@@ -3,5 +3,5 @@ package org.ikuven.bbut.tracking.repository;
 import org.ikuven.bbut.tracking.participant.Participant;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ParticipantRepository extends MongoRepository<Participant, Integer> {
+public interface ParticipantRepository extends MongoRepository<Participant, Long> {
 }
